refactor: make Bubble Tea progress display the default and remove old implementation
