--- example_bubbletea_progress.yaml
@@ -1,69 +1,0 @@
-# Example DAG to test the new Bubble Tea progress display
-# Run with: DAGU_USE_BUBBLETEA_PROGRESS=true dagu start example_bubbletea_progress.yaml
-
-name: bubbletea-progress-demo
-description: Demo of the new Bubble Tea progress display
-
-steps:
-  - name: initialize
-    command: |
-      echo "Starting workflow..."
-      sleep 2
-
-  - name: fetch-data
-    command: |
-      echo "Fetching data from API..."
-      sleep 3
-      echo "Data fetched successfully"
-    depends:
-      - initialize
-
-  - name: process-batch-1
-    command: |
-      echo "Processing batch 1..."
-      sleep 4
-      echo "Batch 1 completed"
-    depends:
-      - fetch-data
-
-  - name: process-batch-2
-    command: |
-      echo "Processing batch 2..."
-      sleep 4
-      echo "Batch 2 completed"
-    depends:
-      - fetch-data
-
-  - name: process-batch-3
-    command: |
-      echo "Processing batch 3..."
-      sleep 4
-      echo "Batch 3 completed"
-    depends:
-      - fetch-data
-
-  - name: aggregate-results
-    command: |
-      echo "Aggregating results from all batches..."
-      sleep 2
-      echo "Aggregation complete"
-    depends:
-      - process-batch-1
-      - process-batch-2
-      - process-batch-3
-
-  - name: generate-report
-    command: |
-      echo "Generating final report..."
-      sleep 3
-      echo "Report generated successfully"
-    depends:
-      - aggregate-results
-
-  - name: notify-completion
-    command: |
-      echo "Sending notification..."
-      sleep 1
-      echo "Workflow completed successfully!"
-    depends:
-      - generate-report
