+++ internal/agent/agent_progress.go
@@ -1,25 +1,14 @@
 package agent
 
 import (
-	"os"
 	"strings"
 
 	"github.com/dagu-org/dagu/internal/digraph"
 )
 
-// createProgressReporter creates the appropriate progress reporter based on configuration
+// createProgressReporter creates the progress reporter
 func createProgressReporter(dag *digraph.DAG, dagRunID string, params []string) ProgressReporter {
-	// Check if we should use the new Bubble Tea implementation
-	useBubbleTea := os.Getenv("DAGU_USE_BUBBLETEA_PROGRESS") == "true"
-
-	if useBubbleTea {
-		display := NewProgressTeaDisplay(dag)
-		display.SetDAGRunInfo(dagRunID, strings.Join(params, " "))
-		return display
-	}
-
-	// Default to the original implementation
-	display := NewProgressDisplay(os.Stderr, dag)
+	display := NewProgressTeaDisplay(dag)
 	display.SetDAGRunInfo(dagRunID, strings.Join(params, " "))
 	return display
 }
